requirements.txt
streamlit
exifread
pandas
numpy
opencv-python
pillow